api/requirements.txt
fastapi==0.119.0
uvicorn==0.38.0
python-multipart==0.0.20
openai-agents==0.4.0
pypdf==5.6.0
python-dotenv==1.1.0
langchain==0.3.27
langchain-community==0.3.29
tiktoken==0.12.0
pydantic==2.12.3
google-auth==2.41.1
google-api-python-client==2.185.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
SQLAlchemy==2.0.44
PyJWT==2.10.1